Started work on powerups
Added classes: PowerUp PowerUpType
Changed Classes: Boat, Lane
added shields to the boats, they initialise with 0

need to figure out how collisions work and add collision detection for the power ups.

have it so that the same number of entities appear on screen but replaced 20% of obstacles with power ups. this can be changed if it's too high or too low it was just a starting value